pyproject.toml
[project]
name = "advice-bot"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.1.0",
    "crcmod>=1.7",
    "discord-py>=2.4.0",
    "flask>=3.0.3",
    "google-cloud-kms>=3.0.0",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.1.0",
    "pytype>=2024.10.11",
    "yapf>=0.40.2",
]

[tool.yapfignore]
ignore_patterns = [
  ".git/",
  ".venv/",
]
